backend/requirements.txt
fastapi
uvicorn
groq
requests
python-dotenv
python-multipart
pydantic
supabase
